pyproject.toml
[project]
name = "aria-voice-agent"
version = "0.1.0"
description = "AI healthcare receptionist — voice agent for appointment scheduling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # NumPy<2 for SciPy/pyloudnorm compatibility (pipecat VAD)
    "numpy<2",
    # Config and validation
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # Voice pipeline
    "pipecat-ai[deepgram,cartesia,openai,silero,webrtc,runner]>=0.0.50",
    "groq>=0.4.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "logfire>=0.1.0",
    "asyncpg>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
